% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_plot.R
\name{path_plot}
\alias{path_plot}
\title{Internal function to render a ggplot with the continuous path}
\usage{
path_plot(df, D)
}
\arguments{
\item{df}{A dataframe}

\item{D}{A vector with two vars}
}
\value{
A ggplot object
}
\description{
Internal function to render a ggplot with the continuous path
}
