% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.yatchew_test}
\alias{print.yatchew_test}
\title{A print method for yatchew_test}
\usage{
\method{print}{yatchew_test}(x, suppress_intro = FALSE, ...)
}
\arguments{
\item{x}{A yatchew_test object}

\item{suppress_intro}{The description of the test is not printed}

\item{...}{Undocumented}
}
\value{
No return, custom print method for yatchew_test objects.
}
\description{
A customized printed display for yatchew_test output
}
