% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.yatchew_test}
\alias{summary.yatchew_test}
\title{A summary method for yatchew_test}
\usage{
\method{summary}{yatchew_test}(object, ...)
}
\arguments{
\item{object}{A yatchew_test object}

\item{...}{Undocumented}
}
\value{
No return, custom summary method for yatchew_test objects.
}
\description{
A customized summary display for yatchew_test output
}
