% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yatchew_test.R
\name{yatchew_test.data.frame}
\alias{yatchew_test.data.frame}
\title{General yatchew_test method for unclassed dataframes}
\usage{
\method{yatchew_test}{data.frame}(data, Y, D, het_robust = FALSE, path_plot = FALSE, ...)
}
\arguments{
\item{data}{(data.frame) A dataframe.}

\item{Y}{(char) Dependent variable.}

\item{D}{(char) Independent variable.}

\item{het_robust}{(logical) If FALSE, the test is performed under the assumption of homoskedasticity, i.e. \eqn{E[V(\varepsilon| X^2)] = \sigma^2}, as in Yatchow (1997). If TRUE, the test is performed using the heteroskedasticity-robust test statistic proposed by de Chaisemartin and D'Haultfoeuille (2024).}

\item{path_plot}{(logical)}

\item{...}{Undocumented.}
}
\value{
A list with test results.
}
\description{
General yatchew_test method for unclassed dataframes
}
