% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yatchew_test.R
\name{yatchew_test}
\alias{yatchew_test}
\title{Main function}
\usage{
yatchew_test(data, ...)
}
\arguments{
\item{data}{A data object.}

\item{...}{Undocumented.}
}
\value{
Method dispatch depending on the data object class.
}
\description{
Test of Linearity of a Conditional Expectation Function (Yatchew, 1997; de Chaisemartin and D'Haultfoeuille, 2024)
}
